consolestigespill interface
