+++ console/src/main/java/no/hvl/dat109/console/SpillControllerImpl.java
@@ -7,6 +7,7 @@ import no.hvl.dat109.spiller.Spiller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+
 @Component
 public class SpillControllerImpl implements SpillController {
